# Conflicts:
#	macosx/EternityEngine.xcodeproj/project.pbxproj
#	macosx/launcher/EternityLaunch.xcodeproj/project.pbxproj
#	source/a_doom.cpp
#	source/c_runcmd.h
#	source/d_main.cpp
#	source/g_cmd.cpp
#	source/g_game.cpp
#	source/m_misc.cpp
#	source/m_vector.h
#	source/p_genlin.cpp
#	source/p_inter.cpp
#	source/p_maputl.cpp
#	source/p_maputl.h
#	source/p_spec.cpp

+++ source/st_hbar.cpp
@@ -77,7 +77,7 @@ void ST_DrawSmallHereticNumber(int val, int x, int y, bool fullscreen)
       if(val > 99999)
          val = 99999;
       sprintf(buf, "%d", val);
-      x -= 4 * (strlen(buf));
+      x -= static_cast<int>(4 * (strlen(buf)));
       for(char *rover = buf; *rover; rover++)
       {
          int i = *rover - '0';
